requirements.txt
aiofile==1.5.2
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
pycups==1.9.74
yarl==1.3.0
